+++ CHANGELOG.md
@@ -64,8 +64,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   core plugins
 - HTTP service got a `-redirect` flag which allows for URL redirection via a
   HTTP 307 response
-- Gitter chatbot tutorial.
+- Support for immediate response in HTTP service
 - Daal4py example usage.
+- Gitter chatbot tutorial.
 ### Changed
 - Renamed `-seed` to `-inputs` in `dataflow create` command
 - Renamed configloader/png to configloader/image and added support for loading JPEG and TIFF file formats